+++ src/bika/lims/content/abstractbaseanalysis.py
@@ -584,6 +584,28 @@ AllowManualUncertainty = BooleanField(
     )
 )
 
+RESULT_TYPES = (
+    ("numeric", _("Numeric")),
+    ("string", _("String")),
+    ("text", _("Text")),
+    ("select", _("Selection list")),
+    ("multiselect", _("Multiple selection")),
+    ("multiselect_duplicates", _("Multiple selection (with duplicates)")),
+    ("multichoice", _("Multiple choices")),
+)
+
+# Type of control to be rendered on results entry
+ResultType = StringField(
+    "ResultType",
+    schemata="Result Options",
+    default="numeric",
+    vocabulary=DisplayList(RESULT_TYPES),
+    widget=SelectionWidget(
+        label=_("Result type"),
+        format="select",
+    )
+)
+
 # Results can be selected from a dropdown list.  This prevents the analyst
 # from entering arbitrary values.  Each result must have a ResultValue, which
 # must be a number - it is this number which is interpreted as the actual
@@ -611,25 +633,12 @@ ResultOptions = RecordsField(
     )
 )
 
-RESULT_OPTIONS_TYPES = (
-    ("select", _("Selection list")),
-    ("multiselect", _("Multiple selection")),
-    ("multiselect_duplicates", _("Multiple selection (with duplicates)")),
-    ("multichoice", _("Multiple choices")),
-)
-
+# TODO Remove ResultOptionsType field. It was Replaced by ResultType
 ResultOptionsType = StringField(
     "ResultOptionsType",
-    schemata="Result Options",
-    default="select",
-    vocabulary=DisplayList(RESULT_OPTIONS_TYPES),
-    widget=SelectionWidget(
-        label=_("Control type"),
-        description=_(
-            "Type of control to be displayed on result entry when predefined "
-            "results are set"
-        ),
-        format="select",
+    readonly=True,
+    widget=StringWidget(
+        visible=False,
     )
 )
 
@@ -663,15 +672,12 @@ ResultOptionsSorting = StringField(
 )
 
 # Allow/disallow the capture of text as the result of the analysis
+# TODO Remove StringResult field. It was Replaced by ResultType
 StringResult = BooleanField(
     "StringResult",
-    schemata="Analysis",
-    default=False,
+    readonly=True,
     widget=BooleanWidget(
-        label=_("String result"),
-        description=_(
-            "Enable this option to allow the capture of text as result"
-        )
+        visible=False,
     )
 )
 
@@ -794,6 +800,7 @@ schema = BikaSchema.copy() + Schema((
     Uncertainties,
     PrecisionFromUncertainty,
     AllowManualUncertainty,
+    ResultType,
     ResultOptions,
     ResultOptionsType,
     ResultOptionsSorting,
@@ -1071,3 +1078,23 @@ class AbstractBaseAnalysis(BaseContent):  # TODO BaseContent?  is really needed?
         """
         tat = self.Schema().getField("MaxTimeAllowed").get(self)
         return tat or self.bika_setup.getDefaultTurnaroundTime()
+
+    # TODO Remove. ResultOptionsType field was replaced by ResulType field
+    def getResultOptionsType(self):
+        if self.getStringResult():
+            return "select"
+        return self.getResultType()
+
+    # TODO Remove. ResultOptionsType field was replaced by ResulType field
+    def setResultOptionsType(self, value):
+        self.setResultType(value)
+
+    # TODO Remove. StringResults field was replaced by ResulType field
+    def getStringResult(self):
+        result_type = self.getResultType()
+        return result_type in ["string", "text"]
+
+    # TODO Remove. StringResults field was replaced by ResulType field
+    def setStringResult(self, value):
+        result_type = "string" if bool(value) else "numeric"
+        self.setResultType(result_type)